backend/requirements.txt
# Agricultural Analysis Game - Backend Dependencies

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Geospatial Libraries
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0
folium==0.14.0

# Satellite Data Processing
earthengine-api==0.1.363
sentinelsat==1.2.1
rioxarray==0.15.0
xarray==2023.7.0

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1

# HTTP Requests
requests==2.31.0
urllib3==2.0.4

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Database (if needed)
SQLAlchemy==2.0.19
sqlite3

# Caching
redis==4.6.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Logging
loguru==0.7.0

# Date/Time
python-dateutil==2.8.2

# Mathematical Operations
sympy==1.12

# Weather Data
pyowm==3.3.0

# NASA APIs
nasa-api==0.1.0

# Async Support
aiohttp==3.8.5
asyncio